created empty components files & reorganized them in "global/" or "chat/" folders
